+++ frontend/capstone-frontend/src/DashComponents/ClientSearchForm.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router'
 function ClientSearchForm() {
     const [query, setQuery] = useState('')
     const [results, setResults] = useState([])
-    const [searched, setSearched] = useState(false)
+    const [status, setStatus] = useState('unrun')
     const navigate = useNavigate()
 
     async function handleSearch(event) {
@@ -13,27 +13,31 @@ function ClientSearchForm() {
             return
         }
 
+        setStatus('loading')
+
         try {   
             const res = await fetch(`${import.meta.env.VITE_API_URL}/providers?search=${encodeURIComponent(query)}`)
             
             if(res.ok) {
                 const data = await res.json()
                 setResults(data)
-                setSearched(true)
+                setStatus('success')
             } else {
                 console.error('Failed to search for providers')
+                setResults([])
+                setStatus('error')
             }
         } catch(error) {
             console.error(error)
             setResults([])
-            setSearched(true)
+            setStatus('error')
         }
     }
 
     function handleClear() {
         setQuery('')
         setResults([])
-        setSearched(false)
+        setStatus('unrun')
     }
 
     return(
@@ -49,16 +53,21 @@ function ClientSearchForm() {
                 <button type="button" onClick={handleClear}>Clear</button>
             </form>
 
-            {searched && results.length === 0 && <p>No matches found.</p>}
-
-            <ul>
-                {results.map(provider => (
-                <li key={provider.id}>
-                    {provider.name} – {provider.servicesOffered?.join(', ')}
-                    <button onClick={() => navigate(`/providers/${provider.id}`)}>See More</button>
-                </li>
-                ))}
-            </ul>
+            {status === 'unrun' && <p>Search for providers.</p>}
+            {status === 'loading' && <p>Loading...</p>}
+            {status === 'error' && <p>Something went wrong.</p>}
+            {status === 'success' && results.length === 0 && <p>No matches found.</p>}
+            {status === 'success' && results.length > 0 && (
+                <ul>
+                    {results.map(provider => (
+                    <li key={provider.id}>
+                        {provider.name} – {provider.servicesOffered?.join(', ')}
+                        <button onClick={() => navigate(`/providers/${provider.id}`)}>See More</button>
+                    </li>
+                    ))}
+                </ul>
+            )}  
+                
         </div>
     )
 }
